+++ server/controllers/files.js
@@ -4,9 +4,18 @@ module.exports = {
   getFiles (query) {
     return (query.tags)
       ? files.getByTags(
-        query.page, query.tags, query.sort || 'id', query.namespace || []
+        query.page,
+        query.tags,
+        query.sort || 'id',
+        query.direction || null,
+        query.namespace || []
       )
-      : files.get(query.page, query.sort || 'id', query.namespace || [])
+      : files.get(
+        query.page,
+        query.sort || 'id',
+        query.direction || null,
+        query.namespace || []
+      )
   },
   getTotalFileCount () {
     return files.getTotalCount()