+++ tests/database.js
@@ -199,6 +199,58 @@ test('database: get files', t => {
   )
 })
 
+test('database: get files sorted descending', t => {
+  t.deepEqual(
+    files.get(1, 'id', 'desc'),
+    [
+      {
+        id: 5,
+        mime: 'image/png',
+        size: 6672,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/d2f5788f623cde1f0fb3dc801396fee235c67ed11d9452bfd765f1331587401d',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/d2f5788f623cde1f0fb3dc801396fee235c67ed11d9452bfd765f1331587401d'
+      },
+      {
+        id: 4,
+        mime: 'image/png',
+        size: 6665,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/6c358705afeeeb6b75ba725cba10145ae366b6c36fe79aa99c983d354926af39',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/6c358705afeeeb6b75ba725cba10145ae366b6c36fe79aa99c983d354926af39'
+      },
+      {
+        id: 3,
+        mime: 'image/png',
+        size: 6117,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42'
+      },
+      {
+        id: 2,
+        mime: 'image/png',
+        size: 5779,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55'
+      }
+    ]
+  )
+})
+
 test('database: get files sorted by size', t => {
   t.deepEqual(
     files.get(1, 'size'),
@@ -251,9 +303,113 @@ test('database: get files sorted by size', t => {
   )
 })
 
+test('database: get files sorted by size ascending', t => {
+  t.deepEqual(
+    files.get(1, 'size', 'asc'),
+    [
+      {
+        id: 1,
+        mime: 'image/png',
+        size: 5012,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c'
+      },
+      {
+        id: 2,
+        mime: 'image/png',
+        size: 5779,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55'
+      },
+      {
+        id: 3,
+        mime: 'image/png',
+        size: 6117,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42'
+      },
+      {
+        id: 4,
+        mime: 'image/png',
+        size: 6665,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/6c358705afeeeb6b75ba725cba10145ae366b6c36fe79aa99c983d354926af39',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/6c358705afeeeb6b75ba725cba10145ae366b6c36fe79aa99c983d354926af39'
+      }
+    ]
+  )
+})
+
 test('database: get files sorted by width', t => {
   t.deepEqual(
     files.get(1, 'width'),
+    [
+      {
+        id: 1,
+        mime: 'image/png',
+        size: 5012,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c'
+      },
+      {
+        id: 2,
+        mime: 'image/png',
+        size: 5779,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55'
+      },
+      {
+        id: 3,
+        mime: 'image/png',
+        size: 6117,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42'
+      },
+      {
+        id: 4,
+        mime: 'image/png',
+        size: 6665,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/6c358705afeeeb6b75ba725cba10145ae366b6c36fe79aa99c983d354926af39',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/6c358705afeeeb6b75ba725cba10145ae366b6c36fe79aa99c983d354926af39'
+      }
+    ]
+  )
+})
+
+test('database: get files sorted by width descending', t => {
+  t.deepEqual(
+    files.get(1, 'width', 'desc'),
     [
       {
         id: 1,
@@ -355,9 +511,165 @@ test('database: get files sorted by height', t => {
   )
 })
 
+test('database: get files sorted by height ascending', t => {
+  t.deepEqual(
+    files.get(1, 'height', 'asc'),
+    [
+      {
+        id: 1,
+        mime: 'image/png',
+        size: 5012,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c'
+      },
+      {
+        id: 2,
+        mime: 'image/png',
+        size: 5779,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55'
+      },
+      {
+        id: 3,
+        mime: 'image/png',
+        size: 6117,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42'
+      },
+      {
+        id: 4,
+        mime: 'image/png',
+        size: 6665,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/6c358705afeeeb6b75ba725cba10145ae366b6c36fe79aa99c983d354926af39',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/6c358705afeeeb6b75ba725cba10145ae366b6c36fe79aa99c983d354926af39'
+      }
+    ]
+  )
+})
+
+test('database: get files sorted by mime', t => {
+  t.deepEqual(
+    files.get(1, 'mime'),
+    [
+      {
+        id: 1,
+        mime: 'image/png',
+        size: 5012,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c'
+      },
+      {
+        id: 2,
+        mime: 'image/png',
+        size: 5779,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55'
+      },
+      {
+        id: 3,
+        mime: 'image/png',
+        size: 6117,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42'
+      },
+      {
+        id: 4,
+        mime: 'image/png',
+        size: 6665,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/6c358705afeeeb6b75ba725cba10145ae366b6c36fe79aa99c983d354926af39',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/6c358705afeeeb6b75ba725cba10145ae366b6c36fe79aa99c983d354926af39'
+      }
+    ]
+  )
+})
+
+test('database: get files sorted by mime descending', t => {
+  t.deepEqual(
+    files.get(1, 'mime', 'desc'),
+    [
+      {
+        id: 1,
+        mime: 'image/png',
+        size: 5012,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c'
+      },
+      {
+        id: 2,
+        mime: 'image/png',
+        size: 5779,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55'
+      },
+      {
+        id: 3,
+        mime: 'image/png',
+        size: 6117,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42'
+      },
+      {
+        id: 4,
+        mime: 'image/png',
+        size: 6665,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/6c358705afeeeb6b75ba725cba10145ae366b6c36fe79aa99c983d354926af39',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/6c358705afeeeb6b75ba725cba10145ae366b6c36fe79aa99c983d354926af39'
+      }
+    ]
+  )
+})
+
 test('database: get files sorted by namespace', t => {
   t.deepEqual(
-    files.get(1, 'namespace', ['namespace']),
+    files.get(1, 'namespace', null, ['namespace']),
     [
       {
         id: 5,
@@ -407,9 +719,61 @@ test('database: get files sorted by namespace', t => {
   )
 })
 
+test('database: get files sorted by namespace descending', t => {
+  t.deepEqual(
+    files.get(1, 'namespace', 'desc', ['namespace']),
+    [
+      {
+        id: 1,
+        mime: 'image/png',
+        size: 5012,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c'
+      },
+      {
+        id: 2,
+        mime: 'image/png',
+        size: 5779,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55'
+      },
+      {
+        id: 3,
+        mime: 'image/png',
+        size: 6117,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42'
+      },
+      {
+        id: 4,
+        mime: 'image/png',
+        size: 6665,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/6c358705afeeeb6b75ba725cba10145ae366b6c36fe79aa99c983d354926af39',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/6c358705afeeeb6b75ba725cba10145ae366b6c36fe79aa99c983d354926af39'
+      }
+    ]
+  )
+})
+
 test('database: get files sorted by invalid namespace', t => {
   t.deepEqual(
-    files.get(1, 'namespace', ['invalid']),
+    files.get(1, 'namespace', null, ['invalid']),
     [
       {
         id: 1,
@@ -476,6 +840,23 @@ test('database: get files by tags', t => {
   )
 })
 
+test('database: get files by tags sorted descending', t => {
+  t.deepEqual(
+    files.getByTags(1, ['lorem', 'ipsum', 'dolor', 'sit', 'amet'], 'desc'),
+    [{
+      id: 1,
+      mime: 'image/png',
+      size: 5012,
+      width: 500,
+      height: 500,
+      mediaUrl: originalBaseUrl +
+        '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c',
+      thumbnailUrl: thumbnailsBaseUrl +
+        '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c'
+    }]
+  )
+})
+
 test('database: get files by tags sorted by size', t => {
   t.deepEqual(
     files.getByTags(1, ['lorem', 'ipsum', 'dolor'], 'size'),
@@ -517,9 +898,91 @@ test('database: get files by tags sorted by size', t => {
   )
 })
 
+test('database: get files by tags sorted by size ascending', t => {
+  t.deepEqual(
+    files.getByTags(1, ['lorem', 'ipsum', 'dolor'], 'size', 'asc'),
+    [
+      {
+        id: 1,
+        mime: 'image/png',
+        size: 5012,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c'
+      },
+      {
+        id: 2,
+        mime: 'image/png',
+        size: 5779,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55'
+      },
+      {
+        id: 3,
+        mime: 'image/png',
+        size: 6117,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42'
+      }
+    ]
+  )
+})
+
 test('database: get files by tags sorted by width', t => {
   t.deepEqual(
     files.getByTags(1, ['lorem', 'ipsum', 'dolor'], 'width'),
+    [
+      {
+        id: 1,
+        mime: 'image/png',
+        size: 5012,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c'
+      },
+      {
+        id: 2,
+        mime: 'image/png',
+        size: 5779,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55'
+      },
+      {
+        id: 3,
+        mime: 'image/png',
+        size: 6117,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42'
+      }
+    ]
+  )
+})
+
+test('database: get files by tags sorted by width ascending', t => {
+  t.deepEqual(
+    files.getByTags(1, ['lorem', 'ipsum', 'dolor'], 'width', 'asc'),
     [
       {
         id: 1,
@@ -599,10 +1062,133 @@ test('database: get files by tags sorted by height', t => {
   )
 })
 
+test('database: get files by tags sorted by height descending', t => {
+  t.deepEqual(
+    files.getByTags(1, ['lorem', 'ipsum', 'dolor'], 'height', 'desc'),
+    [
+      {
+        id: 1,
+        mime: 'image/png',
+        size: 5012,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c'
+      },
+      {
+        id: 2,
+        mime: 'image/png',
+        size: 5779,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55'
+      },
+      {
+        id: 3,
+        mime: 'image/png',
+        size: 6117,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42'
+      }
+    ]
+  )
+})
+
+test('database: get files by tags sorted by mime', t => {
+  t.deepEqual(
+    files.getByTags(1, ['lorem', 'ipsum', 'dolor'], 'mime'),
+    [
+      {
+        id: 1,
+        mime: 'image/png',
+        size: 5012,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c'
+      },
+      {
+        id: 2,
+        mime: 'image/png',
+        size: 5779,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55'
+      },
+      {
+        id: 3,
+        mime: 'image/png',
+        size: 6117,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42'
+      }
+    ]
+  )
+})
+
+test('database: get files by tags sorted by mime descending', t => {
+  t.deepEqual(
+    files.getByTags(1, ['lorem', 'ipsum', 'dolor'], 'mime', 'desc'),
+    [
+      {
+        id: 1,
+        mime: 'image/png',
+        size: 5012,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c'
+      },
+      {
+        id: 2,
+        mime: 'image/png',
+        size: 5779,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55'
+      },
+      {
+        id: 3,
+        mime: 'image/png',
+        size: 6117,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42'
+      }
+    ]
+  )
+})
+
 test('database: get files by tags sorted by namespace', t => {
   t.deepEqual(
     files.getByTags(
-      1, ['lorem', 'ipsum', 'dolor'], 'namespace', ['namespace']
+      1, ['lorem', 'ipsum', 'dolor'], 'namespace', null, ['namespace']
     ),
     [
       {
@@ -642,10 +1228,53 @@ test('database: get files by tags sorted by namespace', t => {
   )
 })
 
+test('database: get files by tags sorted by namespace descending', t => {
+  t.deepEqual(
+    files.getByTags(
+      1, ['lorem', 'ipsum', 'dolor'], 'namespace', 'desc', ['namespace']
+    ),
+    [
+      {
+        id: 1,
+        mime: 'image/png',
+        size: 5012,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/2acedf8e20512a10fc07cceca8d16923e790369b90acebf9efcd926f50dd5c0c'
+      },
+      {
+        id: 2,
+        mime: 'image/png',
+        size: 5779,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/5ef2eac48dd171cf98793df1e123238a61fb8ed766e862042b25467066fabe55'
+      },
+      {
+        id: 3,
+        mime: 'image/png',
+        size: 6117,
+        width: 500,
+        height: 500,
+        mediaUrl: originalBaseUrl +
+          '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42',
+        thumbnailUrl: thumbnailsBaseUrl +
+          '/31426ccc8101461ad30806840b29432fb88bb84687ef9e002976551c8aa08e42'
+      }
+    ]
+  )
+})
+
 test('database: get files by tags sorted by invalid namespace', t => {
   t.deepEqual(
     files.getByTags(
-      1, ['lorem', 'ipsum', 'dolor'], 'namespace', ['invalid']
+      1, ['lorem', 'ipsum', 'dolor'], 'namespace', null, ['invalid']
     ),
     [
       {