+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
++ Added sorting by `mime`
++ Added option to change sort direction when viewing files
+
 ## [3.1.1] - 2018-08-02
 
 ### Fixed